Update to Vite and adding to non-default workers

+++ frontend/vite.config.js
@@ -2,8 +2,7 @@ import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 import fs from 'fs';
 import path from 'path';
-
-const WRONG_CODE = `import { bpfrpt_proptype_WindowScroller } from "../WindowScroller.js";`;
+// import monacoEditorPlugin from 'vite-plugin-monaco-editor';
 
 export default defineConfig({
     esbuild: {
@@ -21,21 +20,34 @@ export default defineConfig({
         {
             name: 'my:react-virtualized',
             configResolved(config) {
-                const file = path.resolve(path.dirname(config.configFile), 'node_modules/react-lazylog/node_modules/react-virtualized/dist/es/WindowScroller/utils/onScroll.js');
-                const code = fs.readFileSync(file, 'utf-8');
-                const modified = code.replace(WRONG_CODE, '');
-                fs.writeFileSync(file, modified);
+                const WRONG_CODE = `import { bpfrpt_proptype_WindowScroller } from "../WindowScroller.js";`;
+                const nodepath = "node_modules/react-virtualized/dist/es/WindowScroller/utils/onScroll.js"
+                // const nodepath = "node_modules/react-virtualized/dist/es/WindowScroller/WindowScroller.js"
+                                 
+                // 'node_modules/react-lazylog/node_modules/react-virtualized/dist/es/WindowScroller/utils/onScroll.js'
+                const file = path.resolve(path.dirname(config.configFile), nodepath);
+                if(fs.existsSync(file) === true) {
+                    console.log('Removing wrong code from react-virtualized')
+                    const code = fs.readFileSync(file, 'utf-8');
+                    const modified = code.replace(WRONG_CODE, '');
+                    fs.writeFileSync(file, modified);
+                }
             },
         },
+        // monacoEditorPlugin.default({
+        //     languageWorkers: ['json', 'python', 'shell'],
+        // }),
     ],
-    define: {
-        'process.env': {},
-    },
+    envDir: './src/enviroment',
+    // define: {
+    //     'process.env': {},
+    // },
     server: {
         port: 3000,
         host: true,
     },
     build: {
-        outDir: './build',
+        outDir: '../app/mainapp/frontbuild',
     },
+    base: '/webapp/',
 });
